move SSO config into env.yml and separate envar / sso stuff from suite/build stuff in the suite object
This will enable us to have multiple target environments from the same test job, and use a parameter to select
them just like we would for the test suite itself.

I've also moved the env.yml option to an argument (required) in the command line processor, and eliminated the option for sso.yml.

Finally, I've added a new parameter at the environment level that allows override of the Maven goals. This is useful when
the target environment won't support uploading (as when we're testing a static proxy).